added doc comments #64
