+++ webapp/src/http_handlers/index/mod.rs
@@ -1,4 +1,6 @@
-/// GET method handling for "/" route
+/// `GET` method handling for "/" route
 pub mod get;
+/// Template context for `index.html.terra`
+mod index_context;
 
 pub use get::get_index;
