+++ webapp/src/http_handlers/sessions/session.rs
@@ -6,7 +6,14 @@ use std::{hash::{Hash, Hasher}, path::{Path, PathBuf}, sync::{Arc, RwLock}};
 use crate::{config_struct::BackendConfig, filework::new_session_folder};
 use super::sessions_tracker::SessionsTracker;
 
-/// Anonymous session data guard
+/// ## Anonymous session data guard.
+/// ----
+/// Fields:
+/// ---
+/// * `id` - session's identifier;
+/// * `last_connection` - session's last connection date;
+/// * `folder` - a path to a folder, reserved for this session's files;
+/// * `source_path` - a path to the session's source code.
 #[derive(Eq, Clone, Serialize, Deserialize)]
 pub struct Session
 {
@@ -18,7 +25,11 @@ pub struct Session
 
 impl Session
 {
-    /// Create a new Session instance with the given id
+    /// ## Create a new `Session` instance with the given id
+    /// ----
+    /// Args:
+    /// ---
+    /// * `id` - session's id.
     pub fn with_id(id: u128) -> Session
     {
         Session
@@ -30,7 +41,11 @@ impl Session
         }
     }
 
-    /// Session builder method, that sets parent_folder
+    /// `Session` builder method, that sets the `folder` field.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `parent_folder` - a new 'folder' field value.
     pub fn folder(mut self, parent_folder: &Path) -> Session
     {
         self.folder = parent_folder.to_owned();
@@ -38,12 +53,31 @@ impl Session
         self
     }
 
+    /// ## Sets `source_path` field.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `source_path` - a new `source_path` field value.
     pub fn set_source(&mut self, source_path: &Path)
     {
         self.source_path = source_path.to_owned();
     }
 
-    /// Establishes a new session, adds it to the sessions tracker
+    /// ## Establishes a new session.
+    /// 
+    /// The new session is addded to the tracker, and a folder for it is created.
+    /// 
+    /// HTTP session cookie `session_id` is set to the new session's id with `htttp_only` flag set to true.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `cookies` - current HTTP session cookies;
+    /// * `tracker` - `SessionsTracker` to add the new session to;
+    /// * `parent_folder` - a folder, in which this session's folder will be created;
+    /// * `logger` - a logger to log to.
+    /// ----
+    /// ## Returns:
+    /// New session's id, or `None` if a folder for it couldn't be created.
     pub(crate) fn establish_new(
         cookies: &CookieJar<'_>, 
         tracker: &SessionsTracker,
@@ -78,6 +112,12 @@ impl Session
         Some(session_id)
     }
 
+    /// ## Updates session's `last_connection` field to `Utc::now()`.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `tracker` - `SessionsTracker` that holds the session;
+    /// * `session_id` - session's id.
     pub(crate) fn update_session(tracker: &Arc<SessionsTracker>, session_id: &u128)
     {
         tracker.set_last_connection(session_id, Utc::now());
