+++ webapp/src/http_handlers/sessions/sessions_tracker.rs
@@ -8,9 +8,13 @@ use serde::{Deserialize, Serialize};
 use crate::filework::delete_folder;
 use super::session::Session;
 
-/// Anonymous user sessions tracker
+/// ## Anonymous user sessions tracker.
+/// _Internal sessions data is wrapped in a mutex, so it is save to use between threads as is._
 /// 
-/// Internal sessions data is wrapped in a mutex, so it is save to use between threads as is
+/// ------
+/// Fields:
+/// * `sessions` - tracked sessions. They are not guaranteed to not be expired at all times;
+/// * `life_duration` - life span of a session, after which it is considered expired.
 #[derive(Serialize, Deserialize)]
 pub struct SessionsTracker 
 {
@@ -20,7 +24,7 @@ pub struct SessionsTracker
 
 impl SessionsTracker
 {
-    /// Creates a new, empty SessionsTracker
+    /// Creates a new, empty `SessionsTracker`
     pub fn new() -> SessionsTracker
     {
         SessionsTracker
@@ -30,7 +34,17 @@ impl SessionsTracker
         }
     }
 
-    /// Deserializes a SessionsTracker istance from a .json file
+    /// ## Deserializes a `SessionsTracker` istance from a `.json` file.
+    /// 
+    /// ----
+    /// Args:
+    /// ---
+    /// * `path` - path to the file.
+    /// ----
+    /// ## Returns:
+    /// `Some(SessionsTracker)` is deserialization was successful, `None` otherwise.  
+    /// 
+    /// `None` may be returned if the file couldn't be read, or its contents could not be deserialized.
     pub fn from_file(path: &Path) -> Option<SessionsTracker>
     {
         let contents;
@@ -47,7 +61,16 @@ impl SessionsTracker
         }
     }
 
-    /// Serializes SessionsTracker into a .json file
+    /// ## Serializes `SessionsTracker` into a `.json` file.
+    /// 
+    /// ----
+    /// Args:
+    /// ---
+    /// * `path` - path to the file.
+    /// ----
+    /// ## Returns:
+    /// * `true`, if serialization was successful;
+    /// * `false`, if the object could not be serialized to text, or the file could not be written to.
     pub fn save(&self, path: &Path) -> bool
     {
         let serialized;
@@ -64,18 +87,32 @@ impl SessionsTracker
         }
     }
 
-    /// Builder method, that sets sessions life duration
+    /// ## Builder method, that sets sessions life duration.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `duration` - new `life_duration` value.
     pub fn life_duration(mut self, duration: &Duration) -> Self
     {
         self.life_duration = duration.to_owned();
         self
     }
 
+    /// ## Locks `sessions_data`.
+    /// ----
+    /// ## Returns:
+    /// `MutexGuard` for `sessions`. If the mutex is poisoned - it will end the process with the exit code of 1.
     fn lock(&self) -> MutexGuard<HashMap<u128, Session>>
     {
         self.sessions.lock().unwrap_or_else(|_| std::process::exit(1))
     }
     
+    /// ## Sets session's `source_path` field.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `session_id` - id of the session to modify;
+    /// * `sourse_path` - new field value.
     pub fn set_source_file(&self, session_id: &u128, source_path: &Path)
     {
         if let Some(session) = self.lock().get_mut(session_id)
@@ -84,6 +121,12 @@ impl SessionsTracker
         }
     }
 
+    /// ## Sets session's `last_connection` field.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `session_id` - id of the session to modify;
+    /// * `new_date` - new field value.
     pub fn set_last_connection(&self, session_id: &u128, new_date: DateTime<Utc>)
     {
         let sessions = &mut self.lock();
@@ -93,6 +136,14 @@ impl SessionsTracker
         }
     }
 
+    /// ## Returns a cloned `Session` object from `sessions`.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `session_id` - id the session to return.
+    /// ----
+    /// ## Returns:
+    /// `Some<Session>`, if the session is present in the `sessions`, `None` if it is not.
     pub fn get_session(&self, session_id: &u128) -> Option<Session>
     {
         if let Some(session) = self.lock().get(session_id)
@@ -105,13 +156,25 @@ impl SessionsTracker
         }
     }
 
-    /// Adds a new session to be tracked
+    /// ## Adds a new session to be tracked.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `session` - a `Session` to insert.
     pub fn insert_session(&self, session: Session)
     {
         self.lock().insert(session.id, session);
     }
 
-    /// Clears expired sessions
+    /// ## Clears expired sessions.
+    /// 
+    /// Session is expired if `current time - last_connection` is more than `life_duration`.
+    /// 
+    /// Deletes sessions from the `sessions` fields, as well as their respective folder.
+    /// 
+    /// ----
+    /// ## Returns:
+    /// Amount of deleted sessions.
     pub fn delete_old(&self) -> usize
     {
         let now = Utc::now();
