+++ webapp/src/http_handlers/submit/mod.rs
@@ -1,6 +1,6 @@
-/// POST method handling for "/submit" route
+/// `POST` method handling for `"/submit"` route
 pub mod post;
-/// Structs used by the "/submit" endpoint
+/// Structs used by the `"/submit"` route
 mod structs;
 
 pub use post::post_submit;
