+++ webapp/src/http_handlers/submit/structs/submit_input.rs
@@ -1,6 +1,12 @@
 use serde::Deserialize;
 
-/// Request body of POST /submit
+/// ## Request body of `POST /submit`.
+/// ----
+/// Fields:
+/// ---
+/// * `lang` - name of the source code language;
+/// * `options` - compilation flags;
+/// * `code` - source code.
 #[derive(Deserialize)]
 pub struct SubmitInput
 {
