+++ webapp/src/http_handlers/submit/structs/submit_output.rs
@@ -1,7 +1,17 @@
 use super::RunnerOutput;
 use serde::Serialize;
 
-/// Response body of POST /submit
+/// ## Response body of `POST /submit`
+/// ----
+/// Fields:
+/// ---
+/// _Compilatin result._
+/// * `status_code` - status code of the compilation;
+/// * `stdout` - output stream data of the compilation;
+/// * `stderr` - error stream data of the compilation;
+/// ----
+/// _Runner result._
+/// * `runner_output` - results of code execution. Can be ommited, in which case it will not be serialized.
 #[derive(Serialize)]
 pub struct SubmitOutput
 {
@@ -14,7 +24,7 @@ pub struct SubmitOutput
 
 impl SubmitOutput
 {
-    /// Create a new instance with compilation results
+    /// Creates a new instance with compilation results.
     pub fn new(status_code: i32, stdout: &str, stderr: &str) -> SubmitOutput
     {
         SubmitOutput {
@@ -25,7 +35,11 @@ impl SubmitOutput
         }
     }
 
-    /// Create a new instance from compilation results
+    /// ## Creates a new instance from compilation results
+    /// ----
+    /// Args:
+    /// ---
+    /// * `data` - output structure, returned by the compiler.
     pub fn from_compiler_result(data: &compiler::data::output_data::OutputData)
         -> SubmitOutput
     {
@@ -38,8 +52,12 @@ impl SubmitOutput
         }
     }
 
-    /// Set runner_output field of the struct
-    pub fn append_runner_output(&mut self, runner_out: runner::data::output::OutputData)
+    /// ## Sets `runner_output` field of the struct.
+    /// ----
+    /// Args:
+    /// ---
+    /// * `runner_out` - output structure, returned by the runner.
+    pub fn set_runner_output(&mut self, runner_out: runner::data::output::OutputData)
         -> &mut Self
     {
         self.runner_output = Some(RunnerOutput::from_runner_result(runner_out));
